todo_api/pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2-binary = "^2.9.9"
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
dev = "todo_api.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
